Ui Updates

+++ app/src/main/java/com/dss/dsboxplus/data/repo/response/AddEstimateRequest.kt
@@ -65,7 +65,7 @@ data class AddEstimateRequest(
 	val lengthCmField: Int? = 0,
 
 	@field:SerializedName("f1ff")
-	var f1ff: Int? = 0,
+	var f1ff: Float? = 0.0F,
 
 
 	@field:SerializedName("businessid")
